[#22] refactor: 클래스 네이밍 수정
