+++ src/main/java/com/fc5/adminback/domain/annual/AnnualResponseDto.java
@@ -1,17 +1,10 @@
 package com.fc5.adminback.domain.annual;
 
+import com.fc5.adminback.domain.member.MemberListResponseDto;
 import com.fc5.adminback.domain.model.Annual;
-import com.fc5.adminback.domain.model.Member;
-import com.fc5.adminback.domain.model.Status;
 import lombok.*;
-import org.hibernate.annotations.ColumnDefault;
-import org.hibernate.annotations.OnDelete;
-import org.hibernate.annotations.OnDeleteAction;
 
-import javax.persistence.*;
-import java.sql.Date;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 
 @Getter
 @Builder
@@ -19,7 +12,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
 public class AnnualResponseDto {
     private Long id;
-    private MemberResponseDto member;
+    private MemberListResponseDto member;
     private LocalDate startDate;
     private LocalDate endDate;
     private String status;
@@ -36,7 +29,7 @@ public class AnnualResponseDto {
                 .reason(annual.getReason())
                 .summary(annual.getSummary())
                 .spentDays(annual.getSpentDays())
-                .member(MemberResponseDto.of(annual.getMember()))
+                .member(MemberListResponseDto.of(annual.getMember()))
                 .build();
     }
 }
