+++ src/main/java/com/fc5/adminback/domain/duty/DutyResponseDto.java
@@ -1,6 +1,6 @@
 package com.fc5.adminback.domain.duty;
 
-import com.fc5.adminback.domain.annual.MemberResponseDto;
+import com.fc5.adminback.domain.member.MemberListResponseDto;
 import com.fc5.adminback.domain.model.Duty;
 import lombok.*;
 
@@ -12,7 +12,7 @@ import java.time.LocalDate;
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
 public class DutyResponseDto {
     private Long id;
-    private MemberResponseDto member;
+    private MemberListResponseDto member;
     private LocalDate dutyDate;
     private String status;
     private String reason;
@@ -23,7 +23,7 @@ public class DutyResponseDto {
                 .dutyDate(duty.getDutyDate())
                 .status(duty.getStatus().getDescription())
                 .reason(duty.getReason())
-                .member(MemberResponseDto.of(duty.getMember()))
+                .member(MemberListResponseDto.of(duty.getMember()))
                 .build();
     }
 }
