+++ src/main/java/com/fc5/adminback/domain/member/MemberListResponseDto.java
@@ -1,4 +1,4 @@
-package com.fc5.adminback.domain.annual;
+package com.fc5.adminback.domain.member;
 
 import com.fc5.adminback.domain.model.Member;
 import com.fc5.adminback.domain.model.Position;
@@ -9,15 +9,17 @@ import lombok.*;
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
-public class MemberResponseDto {
+public class MemberListResponseDto {
 
+    private Long id;
     private String name;
     private String email;
     private Position position;
     private int annualCount;
 
-    public static MemberResponseDto of(Member member) {
-        return MemberResponseDto.builder()
+    public static MemberListResponseDto of(Member member) {
+        return MemberListResponseDto.builder()
+                .id(member.getId())
                 .name(member.getName())
                 .email(member.getEmail())
                 .position(member.getPosition())
